+++ lib/src/model/json/userdata.g.dart
@@ -17,8 +17,6 @@ UserData _$UserDataFromJson(Map<String, dynamic> json) => UserData(
       ..appVersion = (json['appVersion'] as num?)?.toInt() ?? 0
       ..avatarCounter = (json['avatarCounter'] as num?)?.toInt() ?? 0
       ..isDeveloper = json['isDeveloper'] as bool? ?? false
-      ..disableVideoCompression =
-          json['disableVideoCompression'] as bool? ?? false
       ..deviceId = (json['deviceId'] as num?)?.toInt() ?? 0
       ..subscriptionPlanIdStore = json['subscriptionPlanIdStore'] as String?
       ..lastImageSend = json['lastImageSend'] == null
@@ -95,7 +93,6 @@ Map<String, dynamic> _$UserDataToJson(UserData instance) => <String, dynamic>{
       'appVersion': instance.appVersion,
       'avatarCounter': instance.avatarCounter,
       'isDeveloper': instance.isDeveloper,
-      'disableVideoCompression': instance.disableVideoCompression,
       'deviceId': instance.deviceId,
       'subscriptionPlan': instance.subscriptionPlan,
       'subscriptionPlanIdStore': instance.subscriptionPlanIdStore,