+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## 0.0.94
+## 0.0.96
 
+Feature: Show link in chat if the saved media file contains one
+Improve: Verification badge for groups 
+Improve: Huge reduction in app size
+Fix: Crash on older devices when compressing a video
 Fix: Problem with decrypting messages fixed
 
 ## 0.0.93